Add files via upload
A simple game developed using pygame module of Python. It displays a grid of 16 cards that are hidden at the beginning of the game. The player has to click on the card to flip it, only two cards can be opened at a time. The goal is to find similar cards and remove them from the grid. The player who succeeds in removing all the cards wins!